Fixes #3906
Initially, the removed block of code was used to stop the chat modal
from popping up altogether, since the /chats page was usable on
mobile. Since the re-design, only the contact list is shown on
mobile, leaving the modal as the main way to communicate. So, this
intercepting code is actually interfering now.